clang-tidy cleanup

+++ hash.h
@@ -1,8 +1,9 @@
 #ifndef DPLANG_HASH_H
 #define DPLANG_HASH_H
 #include <stdint.h>
+#include <stddef.h>
 
 typedef uint32_t hash_t;
 
-hash_t hash_string(const char *key, int length);
+hash_t hash_string(const char *key, size_t length);
 #endif
